// dashboard/src/@types/enrolledCourse.interface.ts
export interface enrolledCourse{
    courseOverview: string;
    createdAt: string;
    duration: string;
    image: string;
    modality: string;
    objectivesTitle: string;
    postTitle: string;
    price: number;
    title: string;
    topic: string;
    updatedAt: string;
    videoUrl: string;
    __v: number;
    _id: string;
  }
  